requirements.txt
###Discord 
discord-buttons-plugin
discord-py-interactions
discord.py
dislash.py


###Other
aiohttp
async-generator
async-timeout
attrs
certifi
cffi
chardet
charset-normalizer
chromedriver
chromedriver-binary
colorama
configparser
crayons
cryptography
h11
idna
multidict
outcome
pycparser
PyNaCl
pyOpenSSL
PySocks
requests
selenium
six
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
webdriver-manager
wsproto
yarl